% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{build_hubs_from_EP}
\alias{build_hubs_from_EP}
\title{build_hubs_from_EP}
\usage{
build_hubs_from_EP(
  EP,
  k_min = 3,
  resolution = 0.5,
  seed = 42,
  quantile_cutoff = 0,
  method = "log_zscore",
  use_leiden = "on"
)
}
\description{
See the vignette and README for usage.
}
