% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{calculate_hub_stability}
\alias{calculate_hub_stability}
\title{calculate_hub_stability}
\usage{
calculate_hub_stability(
  raw_data,
  observed_hubs,
  weight_mode = c("count_only"),
  dist_col = c("D"),
  q_col = "qvalue",
  breaks = c(0, 10000, 25000, 50000, 1e+05, 250000, 5e+05, 1e+06, 2e+06),
  alpha = 0.8,
  sig_cap = 10,
  sig_beta = 0.5,
  B = 10,
  subsample_frac = 0.8,
  k_min = 3,
  resolution = 0.5,
  quantile_cutoff = 0,
  method = "log_zscore",
  seed = 42
)
}
\description{
See the vignette and README for usage.
}
