% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{compute_weights}
\alias{compute_weights}
\title{compute_weights}
\usage{
compute_weights(
  df,
  weight_mode = c("count_only", "sig_only", "distance_only", "count_sig",
    "zscore_residual", "count_sig_plus_dist_linear", "bin_percentile_plus_sig",
    "bin_log_ratio_sig", "bin_log_ratio_sig_normalized", "bin_diff_global",
    "bin_diff_binmax", "bin_percentile"),
  dist_col = c("D"),
  q_col = "qvalue",
  breaks = c(0, 10000, 25000, 50000, 1e+05, 250000, 5e+05, 1e+06, 2e+06),
  alpha = 0.8,
  sig_cap = 10,
  sig_beta = 0.5
)
}
\description{
See the vignette and README for usage.
}
