% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{fast_comembership}
\alias{fast_comembership}
\title{fast_comembership}
\usage{
fast_comembership(
  enhancers,
  weights,
  quantile_cutoff = 0,
  method = c("log_minmax")
)
}
\description{
See the vignette and README for usage.
}
