% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{preprocess_hichip}
\alias{preprocess_hichip}
\title{preprocess_hichip}
\usage{
preprocess_hichip(
  loop_file,
  tss_input = NULL,
  mode = c("all"),
  promoter_window = 0,
  h3k27ac_peak = NULL,
  min_dist = 10000,
  max_dist = 2e+06
)
}
\description{
See the vignette and README for usage.
}
