% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEhub-package.R
\docType{package}
\name{PEhub-package}
\alias{PEhub}
\alias{PEhub-package}
\title{PEhub: promoter-centric detection of multi-way enhancer hubs}
\description{
PEhub reconstructs multi-way enhancer hubs from pairwise chromatin-interaction
data by modelling enhancer-promoter interactions as promoter-anchored weighted
networks, quantifying enhancer co-membership, partitioning enhancers into
candidate hubs via Leiden community detection, and evaluating hubs against a
distance-matched null model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yourname/PEhub}
  \item Report bugs at \url{https://github.com/yourname/PEhub/issues}
}

}
\author{
\strong{Maintainer}: Yidan Sun \email{syidan@wustl.edu}

Authors:
\itemize{
  \item Yidan Sun \email{syidan@wustl.edu}
  \item Jiang Tan \email{jiangtan@wustl.edu}
}

}
\keyword{internal}
